requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
python-dateutil==2.8.2
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
dlib==19.24.2
imutils==0.5.4
click==8.1.7